add lists, pagination, search fetch

+++ lib/comic_vine.rb
@@ -4,12 +4,92 @@ require "net/http"
 module ComicVine
   class Railtie < Rails::Railtie
     config.after_initialize do
-      keyfile = YAML::load(File.open(Rails.root.join('config', 'cv_key.yml')))
-      ComicVine::API.key = keyfile['cvkey']
+      if File.exists? Rails.root.join('config', 'cv_key.yml')
+        keyfile = YAML::load(File.open(Rails.root.join('config', 'cv_key.yml')))
+        ComicVine::API.key = keyfile['cvkey']
+      else
+        ComicVine::API.key = 'no_keyfile_found'
+      end
     end
   end
   
   class CVError < StandardError  
+  end
+  
+  class CVList
+    include Enumerable
+    
+    attr_reader :total_count
+    attr_reader :offset
+    attr_reader :limit
+    attr_reader :cvos
+    
+    def initialize(resp)      
+      @total_count = resp['number_of_total_results']
+      @offset = resp['offset']
+      @limit = resp['limit']
+    end
+    
+    def each
+      @cvos.each { |c| yield c }
+    end
+    
+    def last
+      @cvos.last
+    end
+    
+    protected
+      def update_ivals(new_cvol)
+        @total_count = new_cvol.total_count
+        @offset = new_cvol.offset
+        @limit = new_cvol.limit
+
+        @cvos = new_cvol.cvos
+      end
+  end
+  
+  class CVObjectList < CVList
+    attr_reader :resource
+    
+    def initialize(resp, resc)      
+      super(resp)
+      
+      @resource = resc
+      @cvos = resp['results'].map{ |r| ComicVine::CVObject.new(r)}
+    end
+    
+    def next_page
+      return nil if (@offset + count) == @total_count
+      update_ivals(ComicVine::API.send(@resource, {:limit => @limit, :offset => (@offset + count)}))
+    end
+    
+    def prev_page
+      return nil if @offset == 0
+      update_ivals(ComicVine::API.send(@resource, {:limit => @limit, :offset => (@offset - count)}))
+    end
+  end
+  
+  class CVSearchList < CVList
+    attr_reader :resource
+    attr_reader :query
+    
+    def initialize(resp, resc, query)      
+      super(resp)
+      
+      @resource = resc
+      @query = query
+      @cvos = resp['results'].map{ |r| ComicVine::CVSearchObject.new(r)}
+    end
+    
+    def next_page
+      return nil if (@offset + count) == @total_count
+      update_ivals(ComicVine::API.search(@resource, @query, {:limit => @limit, :offset => (@offset + count)}))
+    end
+    
+    def prev_page
+      return nil if @offset == 0
+      update_ivals(ComicVine::API.search(@resource, @query, {:limit => @limit, :offset => (@offset - count)}))
+    end
   end
   
   class CVObject
@@ -43,6 +123,12 @@ module ComicVine
     end
   end
   
+  class CVSearchObject < CVObject
+    def fetch
+      ComicVine::API.send(@resource_type, @id)
+    end
+  end
+  
   class API
     cattr_accessor :key
     
@@ -67,7 +153,8 @@ module ComicVine
                       :volumes ].freeze
     
     def self.search res, query, opts={}
-      hit_api(build_url("search", opts)+"&resources=#{res}&query=#{query}")
+      resp = hit_api(build_url("search", opts)+"&resources=#{res}&query=#{query}")
+      ComicVine::CVSearchList.new(resp, res, query)
     end
     
     def self.method_missing(method_sym, *arguments, &block)
@@ -82,11 +169,13 @@ module ComicVine
     
     private
       def self.get_list list_type, opts={}
-        hit_api(build_url(list_type, opts))
+        resp = hit_api(build_url(list_type, opts))
+        ComicVine::CVObjectList.new(resp, list_type)
       end
       
       def self.get_item item_type, id, opts={}
-        hit_api(build_url("#{item_type}/#{id}", opts))
+        resp = hit_api(build_url("#{item_type}/#{id}", opts))
+        ComicVine::CVObject.new(resp['results'])
       end
       
       def self.hit_api url
@@ -94,11 +183,7 @@ module ComicVine
         resp = Net::HTTP.get(url)
         presp = JSON.parse(resp)
         raise CVError, presp['error'] unless presp['status_code'] == 1
-        if presp['results'].kind_of?(Array)
-          presp['results'].map{ |r| ComicVine::CVObject.new(r)}
-        else
-          ComicVine::CVObject.new(presp['results'])
-        end
+        presp
       end
       
       def self.build_url action, opts={}
